# backend/operator-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LOCK CITY Operator API",
    "version": "1.0.0",
    "description": "Digital Administrator contract. Supply Bearer credentials separately; never in this file."
  },
  "servers": [
    {
      "url": "https://lock-city-operator-api.onrender.com"
    }
  ],
  "security": [
    {
      "OperatorBearer": []
    }
  ],
  "paths": {
...
    "/api/operator/v1/printful/status": {
      "get": {
        "operationId": "getPrintfulStatus",
        "security": [
          {
            "OperatorBearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintfulStatus"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Writes disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict, idempotency conflict or operation busy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "WooCommerce or verification error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Not configured, unavailable or uncertain outcome",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/operator/v1/printful/templates": {
      "get": {
        "operationId": "getPrintfulTemplates",
        "security": [
          {
            "OperatorBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintfulTemplateList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Writes disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict, idempotency conflict or operation busy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "WooCommerce or verification error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Not configured, unavailable or uncertain outcome",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/operator/v1/printful/sync-products": {
      "get": {
        "operationId": "getPrintfulSyncProducts",
        "security": [
          {
            "OperatorBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintfulSyncProductList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Writes disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict, idempotency conflict or operation busy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "WooCommerce or verification error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Not configured, unavailable or uncertain outcome",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "OperatorBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "operation_id": {
            "type": "string",
            "format": "uuid"
          },
          "verified": {
            "type": "boolean"
          }
        }
      },
...
      "PrintfulStatus": {
        "type": "object",
        "required": [
          "ok",
          "productTemplatesRead",
          "syncProductsRead"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "productTemplatesRead": {
            "type": "boolean"
          },
          "syncProductsRead": {
            "type": "boolean"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "PrintfulTemplate": {
        "type": "object",
        "required": [
          "id",
          "title",
          "catalogProductId",
          "externalProductId",
          "availableVariantIds",
          "colors",
          "sizes",
          "mockupUrl",
          "placements",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "catalogProductId": {
            "type": [
              "integer",
              "null"
            ]
          },
          "externalProductId": {
            "type": [
              "string",
              "null"
            ]
          },
          "availableVariantIds": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "colors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "sizes": {
            "type": "array",
            "items": {}
          },
          "mockupUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "placements": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "createdAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "PrintfulTemplateList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrintfulTemplate"
            }
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": [
              "integer",
              "null"
            ]
          }
        }
      },
      "PrintfulSyncProduct": {
        "type": "object",
        "required": [
          "syncProductId",
          "externalId",
          "name",
          "variantCount",
          "syncedCount",
          "thumbnailUrl",
          "ignored",
          "variants"
        ],
        "properties": {
          "syncProductId": {
            "type": "integer"
          },
          "externalId": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "variantCount": {
            "type": "integer"
          },
          "syncedCount": {
            "type": [
              "integer",
              "null"
            ]
          },
          "thumbnailUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "ignored": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "variants": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "PrintfulSyncProductList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrintfulSyncProduct"
            }
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": [
              "integer",
              "null"
            ]
          }
        }
      }
    }
  }
}